# Capture non-string metadata