--- a/event_test.go
+++ b/event_test.go
@@ -1,7 +1,6 @@
 package slog
 
 import (
-	"errors"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
@@ -45,26 +44,6 @@ func TestEventRawMetadata(t *testing.T) {
 	assert.Equal(t, map[string]string{
 		"error": assert.AnError.Error(),
 	}, e.Metadata)
-}
-
-func TestEventSpecialErrorCase(t *testing.T) {
-	e := Eventf(ErrorSeverity, nil, "Eaten by a grue", assert.AnError)
-
-	assert.Equal(t, "Eaten by a grue", e.Message)
-	assert.Equal(t, map[string]string{
-		"error": assert.AnError.Error(),
-	}, e.Metadata)
-
-	assert.Equal(t, map[string]interface{}{
-		"error": assert.AnError,
-	}, e.RawMetadata)
-}
-
-func TestEventSpecialErrorCaseWithFormat(t *testing.T) {
-	e := Eventf(ErrorSeverity, nil, "game over: %v", errors.New("eaten by a grue"))
-
-	assert.Equal(t, "game over: eaten by a grue", e.Message)
-	assert.Equal(t, map[string]string(nil), e.Metadata)
 }
 
 type testLogMetadataProvider map[string]string