--- a/event_test.go
+++ b/event_test.go
@@ -29,21 +29,6 @@ func TestEventfMetadataParam(t *testing.T) {
 
 	e := Eventf(CriticalSeverity, nil, "foo: %v", param, metadata)
 	assert.EqualValues(t, metadata, e.Metadata)
-	assert.EqualValues(t, map[string]interface{}{
-		"foo": "foo",
-	}, e.RawMetadata)
-}
-
-func TestEventRawMetadata(t *testing.T) {
-	metadata := map[string]interface{}{
-		"error": assert.AnError,
-	}
-	e := Eventf(CriticalSeverity, nil, "msg", metadata)
-
-	assert.Equal(t, metadata, e.RawMetadata)
-	assert.Equal(t, map[string]string{
-		"error": assert.AnError.Error(),
-	}, e.Metadata)
 }
 
 type testLogMetadataProvider map[string]string
@@ -59,4 +44,28 @@ func TestEventfLogMetadataProvider(t *testing.T) {
 
 	e := Eventf(CriticalSeverity, nil, "foo: %v", param)
 	assert.EqualValues(t, param, e.Metadata)
+}
+
+func BenchmarkLogMetadataInterface(b *testing.B) {
+	for i := 0; i < b.N; i++ {
+		Eventf(ErrorSeverity, nil, "foo", map[string]interface{}{
+			"string": "foo",
+			"number": 42,
+		})
+	}
+}
+
+func BenchmarkLogMetadataStrings(b *testing.B) {
+	for i := 0; i < b.N; i++ {
+		Eventf(ErrorSeverity, nil, "foo", map[string]string{
+			"string": "foo",
+			"number": "42",
+		})
+	}
+}
+
+func BenchmarkLogMetadataInterpolated(b *testing.B) {
+	for i := 0; i < b.N; i++ {
+		Eventf(ErrorSeverity, nil, "foo %s %d", "foo", 42)
+	}
 }