--- a/event_test.go
+++ b/event_test.go
@@ -28,7 +28,22 @@ func TestEventfMetadataParam(t *testing.T) {
 	}
 
 	e := Eventf(CriticalSeverity, nil, "foo: %v", param, metadata)
-	assert.EqualValues(t, metadata, e.Metadata)
+	expected := map[string]interface{}{
+		"foo": "foo",
+	}
+	assert.EqualValues(t, expected, e.Metadata)
+}
+
+func TestEventfMetadataParamInterface(t *testing.T) {
+	metadata := map[string]interface{}{
+		"foo": 3,
+	}
+
+	e := Eventf(CriticalSeverity, nil, "foo", metadata)
+	expected := map[string]interface{}{
+		"foo": 3,
+	}
+	assert.EqualValues(t, expected, e.Metadata)
 }
 
 type testLogMetadataProvider map[string]string
@@ -43,7 +58,10 @@ func TestEventfLogMetadataProvider(t *testing.T) {
 	}
 
 	e := Eventf(CriticalSeverity, nil, "foo: %v", param)
-	assert.EqualValues(t, param, e.Metadata)
+	expected := map[string]interface{}{
+		"foo": "bar",
+	}
+	assert.EqualValues(t, expected, e.Metadata)
 }
 
 func BenchmarkLogMetadataInterface(b *testing.B) {