This change introduces the concept of 'non-string metadata' to slog. There are two main reasons for doing this:

1. We wish to use non-string data in downstream systems. Our two current use cases for this are; sending errors to an error tracking system and indexing non-string metadata natively in elasticsearch (e.g. we want to be able to emit an array of tags).
2. Aligning the interface with expectations. We currently have multiple instances of incorrect internal uses where a `map[string]interface{}` is passed, and this metadata is currently silently dropped.

We choose to add a `RawMetadata` field to maintain backwards compatibility. We provide an invariant that anything in the `Metadata` map will be the equivalent stringified version of the value in ` RawMetadata`, as serialized by Go's native `Sprint` function.

The downsides to this are that we increase the memory footprint of a `slog.Event` and the allocations performed to log something. However, there are probably many optimisations that we can do to alleviate this if it becomes an issue. Here are some simple benchmarks for logging with 2 params.

Before
```
➜  slog git:(master) ✗ go test -bench . -benchmem  .
goos: darwin
goarch: amd64
pkg: github.com/monzo/slog
BenchmarkLogMetadataStrings-8        	 1000000	      1027 ns/op	     905 B/op	      11 allocs/op
BenchmarkLogMetadataInterpolated-8   	  766807	      1575 ns/op	     743 B/op	      13 allocs/op
PASS
```

After
```
➜  slog git:(will-slog-metadata) ✗ go test -bench . -benchmem  .
goos: darwin
goarch: amd64
pkg: github.com/monzo/slog
BenchmarkLogMetadataInterface-8      	  696866	      1439 ns/op	    1251 B/op	      15 allocs/op
BenchmarkLogMetadataStrings-8        	  897459	      1324 ns/op	    1277 B/op	      15 allocs/op
BenchmarkLogMetadataInterpolated-8   	  667189	      1621 ns/op	     743 B/op	      13 allocs/op
PASS
```