@Willyham 2020-06-05
> Broadly, this looks good to me. I'm not sure I'm totally sold on keeping the existing `Metadata` around _and_ the `RawMetadata` though. I think this will be quite confusing for people in the future. Instead, I wonder if we should just change the type of `Metadata` to `map[string]interface{}` and update anywhere internally we expect it to be the old type?

Are we happy with introducing a breaking change here? This library is open source. 

I think the only way we can change the Metadata type and not break things for external consumers is to update all of our internal usages to a new path (either a closed-source fork or a versioned package).