@obeattie 2020-06-05
>Are we happy with introducing a breaking change here? This library is open source.

Personally, I am fine with that for multiple reasons:

* There's no published guarantee that the interfaces won't change
* This is a pretty minor change, and I think for _most_ code wouldn't even cause a type warning
* I'd honestly be a bit surprised if this is very used widely outside Monzo