@Willyham 2020-06-05
ok, I've updated this to go with the simpler, breaking, implementation. I'm going to apply this locally and ensure that its a feasible internal upgrade across the board, then come back to confirm.